\name{federalregister-package}
\alias{federalregister-package}
\docType{package}
\title{Client package for the U.S. Federal Register API}
\description{Access United States Federal Register data from R.}
%\details{https://www.federalregister.gov/blog/learn/developers}
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}

