\name{fr_search}
\alias{fr_search}
\title{Search the Federal Register}
\description{Function to search for entries in the Federal Register}
\usage{fr_search()}
\arguments{
  \item{}{}
  \item{version}{The Federal Register API version. Default is 'v1'.}
}
\details{}

%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
\examples{
\dontrun{fr_search()}
}
%\keyword{}
