\name{pi_get}
\alias{pi_get}
\title{Get Public Inspection Metadata}
\description{Function to get metadata for one or more Public Inspection Documents}
\usage{}
\arguments{
  \item{}{}
  \item{...}{}
}
\details{}

%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
\examples{}
%\keyword{}
