\name{pi_search}
\alias{pi_search}
\title{Search Public Inspection Documents}
\description{Function to search for documents under Public Inspection}
\usage{pi_search()}
\arguments{
  \item{}{}
  \item{version}{The Federal Register API version. Default is 'v1'.}
}
%\details{}
\value{A list containing one or more objects of class \code{fedreg_document}.}
\references{
\url{https://www.federalregister.gov/learn/public-inspection-desk-2}
}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
\examples{
\dontrun{pi_search()}
}
%\keyword{}
