\name{fr_agencies}
\alias{fr_agencies}
\alias{print.fedreg_agency}
\title{Get Federal Register Agency Information}
\description{Function to get information about a single federal agency listed in the Federal Register}
\usage{}
\arguments{
  \item{}{}
  \item{...}{}
}
\details{}

%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
\examples{}
%\keyword{}
