\name{fr_get}
\alias{fr_get}
\title{Get Federal Register Metadata}
\description{Function to get metadata for one or more Federal Register Documents}
\usage{}
\arguments{
  \item{}{}
  \item{...}{}
}
\details{}

%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
\examples{}
%\keyword{}
