\name{pi_search}
\alias{pi_search}
\title{Search Public Inspection Documents}
\description{Function to search for documents under Public Inspection}
\usage{}
\arguments{
  \item{}{}
  \item{...}{}
}
\details{}

%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
\examples{}
%\keyword{}
